requirements.txt
Crawl4AI==0.4.247
python-dotenv==1.0.1
pydantic==2.10.6
flask==2.3.3
pymongo==4.6.1
asyncio==3.4.3
requests==2.31.0
gunicorn